requirements.txt
python-dotenv>=1.0.1
httpx>=0.28.1
mcp[cli]>=1.2.1
starlette>=0.45.3
uvicorn>=0.34.0
